# public/apidoc.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Arachne
  license:
    name: MIT
servers:
  - url: /
paths:
...
  /login:
    post:
      summary: Authentificate the user and returns a json web token.
      operationId: auth
      tags:
        - auth
      security: []
      requestBody:
        description: Login credentials.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_name:
                  type: string
                password:
                  type: string
              required:
                - user_name
                - password
      responses:
        200:
          description: Login information.
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                  message:
                    type: message
                  token:
                    type: string
                    example: "eya4az6d4.Azdad8azd.Azdazdf851"
  /users:
    get:
      summary: Get list of users.
      operationId: getUsers
      tags:
        - users
      responses:
        '200':
          description: A list of users.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArrayOfUsers'
    post:
      summary: Create a user.
      operationId: createUser
      description: Requires CREATE_USER permission.
      tags:
        - users
      requestBody:
        description: User to create. Password may be null if external auth service is used.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_name:
                  type: string
                password:
                  type: string
              required:
                - user_name
      responses:
        '200':
          description: User created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /skills:
    get:
      summary: Get skills of brain.
      description: Requires SEE_SKILLS permission (and SEE_SKILLS_ANY to see private skills).
      operationId: reloadBrain
      tags:
        - skills
      responses:
        '200':
          description: List of accessible skills.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfSkills'
    put:
      summary: Create a new skill.
      description: Requires CREATE_SKILL permission. Skill code will be validated against template.
      operationId: createSkill
      tags:
        - skills
      requestBody:
        description: Skill to save.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                skill_name:
                  type: string
                skill_code:
                  type: string
                skill_secret:
                  type: string
              required:
                - skill_name
                - skill_code
      responses:
        '200':
          description: Skill created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /skills/{name}:
    summary: Represents a skill.
    description: Skills have a unique name.
    delete:
      summary: Delete a skill
      operationId: deleteSkill
      description: Requires DELETE_SKILL permision (cf DELETE_SKILL_ANY).
      tags:
        - skills
      parameters:
        - name: name
          in: path
          description: Skill unique name.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Skill deleted.
          content:
            application/json:
              $ref: '#/components/schemas/ApiResponse'
...
  /skills/{name}/edit:
    get:
      summary: Get skill code.
      operationId: getSkillCode
      description: Requires SEE_SKILL_CODE (cf SEE_SKILL_CODE_ANY).
      tags:
        - skills
      parameters:
        - name: name
          in: path
          description: Skill unique name.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Skill reloaded.
          content:
            application/json:
              type: object
              properties:
                success:
                  type: boolean
                message:
                  type: string
                code:
                  type: string
...
  /connectors:
    get:
      summary: Get list of connectors.
      operationId: getConnectors
      description: Requires SEE_ADAPTERS (cf SEE_ADAPTERS_ANY).
      tags:
        - connectors
      responses:
        '200':
          description: List of adapters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfConnectors'
    put:
      summary: Create a new connector.
      operationId: createConnectors
      description: Requires CREATE_ADAPTER.
      tags:
        - connectors
      requestBody:
        description: Connector to create.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: RocketChat-Adapter
                address:
                  type: string
                  example: "127.0.0.1"
              required:
                - name
      responses:
        '200':
          description: Adapter created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "Connector successfully created."
                  connector:
                    type: object
                    $ref: '#/components/schemas/ConnectorDetails'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          examples:
            - "Action is a success."
            - "Action failed because..."
    User:
      type: object
      properties:
        id:
          type: string
        user_name:
          type: string
          example: "the.great.master"
        registered_date:
          type: string
          example: "2018-05-23T12:42:48.822Z"
        last_connect:
          type: string
          example: "2018-05-23T12:42:48.822Z"
        roles:
          type: array
          items:
            type: string
          example: [admin]
        permissions:
          type: array
          items:
            type: string
          example: [SEE_SKILLS, CREATE_USER]
      required:
        - id
    ArrayOfUsers:
      type: array
      items:
        $ref: '#/components/schemas/User'
...
    ArrayOfSkills:
      type: object
      properties:
        skillname:
          $ref: '#components/schemas/Skill'
...
    ConnectorDetails:
      type: object
      properties:
        _id:
          type: string
          example: 5b03c914b42d0319c87880de
        name:
          type: string
          example: RocketChat-Connector
        active:
          type: boolean
        ip:
          type: string
          example: "127.0.0.1"
        token:
          type: string
          example: 27e7d4f1355e81638d2ae225caa860c99d424
      required:
        - _id
    ArrayOfConnectors:
      type: array
      items:
        $ref: '#components/schemas/Connector'
...
security:
  - bearerAuth: []